Fix to remedy image link problem.

+++ dao/src/main/java/com/inthinc/pro/model/VehicleReportItem.java
@@ -6,9 +6,9 @@ public class VehicleReportItem
     private Vehicle vehicle;
     private String makeModelYear;
     private Integer milesDriven;
-    private Integer overallScore;
-    private Integer speedScore;
-    private Integer styleScore;
+    private Float overallScore;
+    private Float speedScore;
+    private Float styleScore;
     private String styleOverall;
     private String styleSpeed;
     private String styleStyle;
@@ -56,27 +56,27 @@ public class VehicleReportItem
     {
         this.milesDriven = milesDriven;
     }
-    public Integer getOverallScore()
+    public Float getOverallScore()
     {
         return overallScore;
     }
-    public void setOverallScore(Integer overallScore)
+    public void setOverallScore(Float overallScore)
     {
         this.overallScore = overallScore;
     }
-    public Integer getSpeedScore()
+    public Float getSpeedScore()
     {
         return speedScore;
     }
-    public void setSpeedScore(Integer speedScore)
+    public void setSpeedScore(Float speedScore)
     {
         this.speedScore = speedScore;
     }
-    public Integer getStyleScore()
+    public Float getStyleScore()
     {
         return styleScore;
     }
-    public void setStyleScore(Integer styleScore)
+    public void setStyleScore(Float styleScore)
     {
         this.styleScore = styleScore;
     }
